perf(chain): rename eth to ethereum
